pyproject.toml
[project]
name = "collab_env"
version = "0.0.1"
description = "Basic scientific computing and deep learning environment with CUDA support"
readme = "README.rst"
requires-python = ">=3.8,<3.11"
license = { text = "Apache-2.0" }
keywords = ["scientific computing", "deep learning", "CUDA", "data processing", "plotting"]
authors = [{ name = "Basis" }]
classifiers = [
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.10"
]

dependencies = [
  "torch>=2.1,<2.4",
  "torchvision",
  "pyro-ppl",
  "numpy<2",
  "pandas",
  "scipy",
  "scikit-learn",
  "opencv-python",
  "tqdm",
  "pyyaml",
  "pyexiftool",
  "matplotlib",
  "seaborn",
  "plotly",
  "ultralytics",
  "dotenv",
  "jupyterlab", 
  "ipywidgets",
  "gcsfs",
  "loguru",
  "pylibjpeg-libjpeg",
  "rfdetr",
  "open3d",
  "gymnasium",
  "torch_geometric",
  "ipympl",
  "pyarrow",
  "pycolmap",
  "pyvista",
  "trame",
  "trame-vtk",
  "trame-vuetify",
  "lpips",
  "scikit-image"
]

[tool.setuptools]                         # NEW
packages = ["collab_env"]  

[project.optional-dependencies]
docs = [
  "setuptools",
  "sphinx",
  "sphinxcontrib-bibtex",
  "sphinx_rtd_theme",
  "myst-parser",
  "nbsphinx",
]
dev = [
  "setuptools",
  "sphinx",
  "sphinxcontrib-bibtex",
  "sphinx_rtd_theme",
  "myst-parser",
  "nbsphinx",
  "pytest",
  "pytest-cov",
  "pytest-xdist",
  "pytest-benchmark",
  "mypy",
  "ruff",
  "nbval",
  "nbqa",
  "black",
  "isort",
  "flake8>=6.0",
]

[build-system]
requires = ["setuptools<70.0"]
build-backend = "setuptools.build_meta"

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "E501"]
exclude = [".venv"]

[tool.nbqa.addopts]
flake8 = [
    "--extend-ignore=E203,E501"
]

[tool.mypy]
ignore_missing_imports = true
warn_unused_ignores = true
exclude = "build"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = ["./tests"]

[tool.isort]
profile = "black"
skip_gitignore = true
